+++ packages/ui-angular/src/lib/components/index.ts
@@ -5,5 +5,6 @@ export type {
   VariableConfig,
 } from '@nubebytes/ui-react';
 export { ButtonComponent } from './button/button.component';
+export { ReactColorPickerComponent } from './react-color-picker/react-color-picker.component';
 export { ReactEditorWrapperComponent } from './react-editor-wrapper/react-editor-wrapper.component';
 export { ReactTestComponent } from './react-test/react-test.component';