+++ packages/ui-react/src/components/editor/components/ui/layer-actions-menu/index.tsx
@@ -3,6 +3,7 @@ import {
   CustomButtonProps,
 } from '@/editor/components/ui/custom-btn';
 import { Action } from '@/editor/hooks/useEditor';
+import { useCanvas } from '@/editor/stores/canvas.store';
 import { getIconElement } from '@/editor/utils/icons.utils';
 import { Button, Menu, MenuHandler, MenuList } from '@material-tailwind/react';
 import { TLUiIconType } from '@tldraw/tldraw';
@@ -10,16 +11,16 @@ import {
   ArrowDownIcon,
   ArrowUpIcon,
   BringToFrontIcon,
+  FlipHorizontal2Icon,
+  FlipVertical2Icon,
   Grid3X3Icon,
+  ImageDownIcon,
+  Layers3Icon,
   LucideIcon,
   MagnetIcon,
   SendToBackIcon,
 } from 'lucide-react';
 import { PropsWithChildren } from 'react';
-
-type LayerActionsProps = {
-  onAction: (action: Action) => void;
-} & Omit<CustomButtonProps, 'children'>;
 
 type LayerAction = {
   action: Action;
@@ -39,6 +40,11 @@ const layerActions: LayerAction[] = [
     icon: BringToFrontIcon,
   },
   {
+    action: 'duplicate',
+    tooltip: 'Duplicar',
+    icon: 'duplicate',
+  },
+  {
     action: 'send-backward',
     tooltip: 'Enviar hacia atrás',
     icon: ArrowDownIcon,
@@ -49,44 +55,87 @@ const layerActions: LayerAction[] = [
     icon: SendToBackIcon,
   },
   {
-    action: 'duplicate',
-    tooltip: 'Duplicar',
-    icon: 'duplicate',
+    action: 'pack',
+    tooltip: 'Amontonar capas en orden',
+    icon: Layers3Icon,
   },
   {
+    action: 'flip-horizontal',
+    tooltip: 'Voltear horizontal',
+    icon: FlipHorizontal2Icon,
+  },
+  {
+    action: 'flip-vertical',
+    tooltip: 'Voltear vertical',
+    icon: FlipVertical2Icon,
+  },
+  {
+    action: 'export-as-png',
+    tooltip: 'Exportar como PNG',
+    icon: ImageDownIcon,
+  },
+];
+
+const genericActions: LayerAction[] = [
+  {
     action: 'toggle-grid',
-    tooltip: 'Mostrar/esconder grid',
+    tooltip: 'Mostrar/esconder cuadrícula',
     icon: Grid3X3Icon,
   },
   {
     action: 'toggle-snap-mode',
-    tooltip: 'Activar/desactivar detección de bordes',
+    tooltip: 'Activar/desactivar bordes inteligentes',
     icon: MagnetIcon,
   },
 ];
 
-export const LayerActionsMenu = ({
+type ActionBtnProps = {
+  disabled?: boolean;
+};
+
+type LayerActionsProps = {
+  onAction: (action: Action) => void;
+} & Omit<CustomButtonProps, 'children'>;
+
+export const ActionsMenu = ({
   onAction,
   children,
   ...props
 }: PropsWithChildren<LayerActionsProps>) => {
+  const { selectedShapes } = useCanvas();
+
+  const actionBtn = ({
+    action,
+    tooltip,
+    icon,
+    disabled,
+  }: LayerAction & ActionBtnProps) => (
+    <CustomButton
+      key={action}
+      tooltip={tooltip}
+      variant="text"
+      size="sm"
+      disabled={disabled}
+      onClick={() => onAction(action)}>
+      {getIconElement(icon)}
+    </CustomButton>
+  );
+
   return (
     <Menu allowHover>
       <MenuHandler>
         <Button {...props}>{children}</Button>
       </MenuHandler>
-      <MenuList className="grid grid-cols-3">
-        {layerActions.map(({ tooltip, action, icon }) => (
-          <div key={action} className="focus:outline-none">
-            <CustomButton
-              tooltip={tooltip}
-              variant="text"
-              size="sm"
-              onClick={() => onAction(action)}>
-              {getIconElement(icon)}
-            </CustomButton>
-          </div>
-        ))}
+      <MenuList>
+        <div className="grid grid-cols-3 focus:outline-none">
+          {layerActions.map((a) =>
+            actionBtn({ ...a, disabled: !selectedShapes.length }),
+          )}
+        </div>
+        <hr className="my-3" />
+        <div className="grid grid-cols-3 focus:outline-none">
+          {genericActions.map(actionBtn)}
+        </div>
       </MenuList>
     </Menu>
   );